+++ crates/nostr/CHANGELOG.md
@@ -23,6 +23,11 @@
 
 -->
 
+## Unreleased
+
+### Added
+- Add nip47 holdinvoice methods and notification
+
 ## v0.43.0 - 2025/07/28
 
 ### Breaking changes